Add hints as to how to pre-allocate entries
Currently whenever a snapshot occurs the Cache is reset and so many
allocations have to take place as the same type of data is re-added to
the Cache.

This commit allows the stores to keep track of the number of values
within an entry and use that size as a hint when the same entry needs to
be recreated after a snapshot.